Lesson 13 - 23:58 - I SURRENDER
I changed all the values from AMOUNT back to PRICE, so that it matches WesBos' example. I guess that's all I can do now. I'll adapt things in the future. When I tried to deploy the new schema to prisma on the back end it kept giving me the error message that nodes had already been created and I couldn't get it to work. Still the placeholder for the Price isn't inputting a value. It used to work, not sure why not now.